+++ src/main/java/de/dm/intellij/liferay/language/groovy/LiferayServerGroovyProgramRunner.java
@@ -1,0 +1,34 @@
+package de.dm.intellij.liferay.language.groovy;
+
+import com.intellij.execution.ExecutionException;
+import com.intellij.execution.configurations.RunProfile;
+import com.intellij.execution.configurations.RunProfileState;
+import com.intellij.execution.configurations.RunnerSettings;
+import com.intellij.execution.executors.DefaultRunExecutor;
+import com.intellij.execution.impl.DefaultJavaProgramRunner;
+import com.intellij.execution.runners.ExecutionEnvironment;
+import com.intellij.execution.runners.ProgramRunner;
+import com.intellij.execution.ui.RunContentDescriptor;
+import com.intellij.openapi.project.Project;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+
+public class LiferayServerGroovyProgramRunner extends DefaultJavaProgramRunner {
+
+    @Override
+    public @NotNull @NonNls String getRunnerId() {
+        return "LiferayServerGroovyProgramRunner";
+    }
+
+    @Override
+    public boolean canRun(@NotNull String executorId, @NotNull RunProfile profile) {
+        return DefaultRunExecutor.EXECUTOR_ID.equals(executorId) && profile instanceof LiferayServerGroovyConfiguration;
+    }
+
+    @Override
+    protected RunContentDescriptor doExecute(@NotNull RunProfileState state, @NotNull ExecutionEnvironment env) throws ExecutionException {
+        RunContentDescriptor runContentDescriptor = super.doExecute(state, env);
+
+        return runContentDescriptor;
+    }
+}