+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Ability to use Freemarker debugger for Modules based on direct service calls to Advanced Resources Importer
 - Feature: JavaScript support for liferay-look-and-feel.xml
 - Feature: Add javax.portlet.version to properties of Portlet components
+- Feature: Run groovy scripts directly on a running Liferay server
 - Change: Update support for IntelliJ 2022.1
 - Bugfix: ClassCastException during code completions in an empty frontend-token-definition.json file
 - Bugfix: description tag inside notification of a workflow definition supports Freemarker, too